% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packpop2-package.R
\docType{package}
\name{packpop2}
\alias{packpop2}
\title{packpop2: A package with a minimum example of package creation.}
\description{
This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/NiloofarNL/packpop2}
}
\author{
Niloofar Nalaee, School of Earth, Environment and Society, McMaster University \email{nalaeen@mcmaster.ca}
}
