% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packpop2-package.R
\docType{data}
\name{Population2016}
\alias{Population2016}
\title{Population of Mashhad in 2016.}
\format{
A data frame with 173 rows and 14 variables:
\describe{
\item{OBJECTID}{Id of each block as a factor}
\item{Neighbourhoods}{The names of Mashhad's neighbourhoods}
\item{Family}{Number of families living in a block}
\item{Owners}{Number of people who has their own property in a block}
\item{Tenants}{Number of people who rent a property in a block}
\item{Total Residential Units}{Whole number of residencial suits in a block}
\item{Apartment Units}{Number of suits in apartments in a block}
\item{House Units}{Number of village houses in a block}
\item{Cottages or sheds}{Number of sheds in a block}
\item{Shape_Length}{Surrounding a block in meter scale}
\item{Shape_Area}{Surface area of a block in meter scale}
\item{Area(Hectare)}{Surface area of a neighbourhoods in hectare scale}
\item{PopulationDensity}{Population density of each neighbourhood}
\item{Population}{Whole population of each neighbourhood}
}
}
\source{
(https://WWW.amar.org.ir)
}
\usage{
data(Population2016)
}
\description{
A dataset containing information about population based on number of families, owners, tenants and residential units of all Neighbourhoods in Mashhad, Iran.
}
\examples{
 data(Population2016)
 GDPPC <- Population2016$GDPPC
}
\keyword{datasets}
