% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_density.R
\name{population_density}
\alias{population_density}
\title{Population density in each block}
\usage{
population_density(a, b)
}
\arguments{
\item{a}{A number of population in a block}

\item{b}{A number of Shape_Area in a block}
}
\value{
The product of \code{a} and \code{b}
}
\description{
Population density in each block
}
\examples{
population_density(96, 9250)
population_density(108, 13246)
}
