% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packpop2-package.R
\docType{package}
\name{packpop2-package}
\alias{packpop2}
\alias{packpop2-package}
\title{packpop2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NiloofarNL/packpop2}
  \item Report bugs at \url{https://github.com/NiloofarNL/packpop2/issues}
}

}
\author{
\strong{Maintainer}: Niloofar Nalaee \email{nalaeen@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
